refactor(workspace): keep resolver outside CLI boundary
